split io package for tee
