+++ io/tee.go
@@ -1,4 +1,4 @@
-package main
+package io
 
 import (
 	"bufio"
